+++ lib/OpenPGP/Packet/ExperimentalPacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet;
+
+/**
+ * OpenPGP Private or Experimental packet (tags 60..63).
+ *
+ * @see http://tools.ietf.org/html/rfc4880#section-4.3
+ */
+class ExperimentalPacket extends Packet
+{
+
+}
