+++ lib/OpenPGP/Packet/MarkerPacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet;
+
+/**
+ * OpenPGP Marker packet (tag 10).
+ *
+ * @see http://tools.ietf.org/html/rfc4880#section-5.8
+ */
+class MarkerPacket extends Packet
+{
+    // @todo
+}
