+++ lib/OpenPGP/Packet/PublicKeyPacket/SecretSubkeyPacket.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace OpenPGP\Packet\PublicKeyPacket;
+
+/**
+ * OpenPGP Secret-Subkey packet (tag 7).
+ *
+ * @see http://tools.ietf.org/html/rfc4880#section-5.5.1.4
+ * @see http://tools.ietf.org/html/rfc4880#section-5.5.3
+ * @see http://tools.ietf.org/html/rfc4880#section-11.2
+ * @see http://tools.ietf.org/html/rfc4880#section-12
+ */
+class SecretSubkeyPacket extends SecretKeyPacket
+{
+    // @todo
+}
