+++ lib/OpenPGP/Packet/SignaturePacket/SubPacket/ExportableCertificationPacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet\SignaturePacket\SubPacket;
+
+class ExportableCertificationPacket extends Subpacket {
+    function read() {
+        $this->data = (ord($this->input) != 0);
+    }
+
+    function body() {
+        return chr($this->data ? 1 : 0);
+    }
+}
