+++ lib/OpenPGP/Packet/SignaturePacket/SubPacket/KeyExpirationTimePacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet\SignaturePacket\SubPacket;
+
+class KeyExpirationTimePacket extends Subpacket {
+    function read() {
+        $this->data = $this->read_timestamp();
+    }
+
+    function body() {
+        return pack('N', $this->data);
+    }
+}
