+++ lib/OpenPGP/Packet/SignaturePacket/SubPacket/KeyFlagsPacket/FeaturesPacket.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace OpenPGP\Packet\SignaturePacket\SubPacket\KeyFlagsPacket;
+
+class FeaturesPacket extends KeyFlagsPacket
+{
+    // Identical functionality to parent
+}
