+++ lib/OpenPGP/Packet/SignaturePacket/SubPacket/PreferredKeyServerPacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet\SignaturePacket\SubPacket;
+
+class PreferredKeyServerPacket extends Subpacket {
+    function read() {
+        $this->data = $this->input;
+    }
+
+    function body() {
+        return $this->data;
+    }
+}
