+++ lib/OpenPGP/Packet/SignaturePacket/SubPacket/SignatureExpirationTimePacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet\SignaturePacket\SubPacket;
+
+class SignatureExpirationTimePacket extends Subpacket {
+    function read() {
+        $this->data = $this->read_timestamp();
+    }
+
+    function body() {
+        return pack('N', $this->data);
+    }
+}
