+++ lib/OpenPGP/Packet/SignaturePacket/SubPacket/SignersUserIDPacket.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OpenPGP\Packet\SignaturePacket\SubPacket;
+
+class SignersUserIDPacket extends Subpacket {
+    function read() {
+        $this->data = $this->input;
+    }
+
+    function body() {
+        return $this->data;
+    }
+}
