+++ lib/OpenPGP/Packet/UserAttributePacket.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace OpenPGP\Packet;
+
+/**
+ * OpenPGP User Attribute packet (tag 17).
+ *
+ * @see http://tools.ietf.org/html/rfc4880#section-5.12
+ * @see http://tools.ietf.org/html/rfc4880#section-11.1
+ */
+class UserAttributePacket extends Packet {
+    public $packets;
+
+    // @todo
+}
